refactor users handlers
